:white_check_mark: Separate the tape and qnode tests
